backend/requirements.txt
flask
pillow
boto3
pymongo
flask-cors
opencv-python
google-generativeai
firebase-admin
certifi